requirements.txt
anyio==3.5.0
apischema==0.16.6
appnope==0.1.2
asttokens==2.0.5
asyncio==3.4.3
attrs==21.4.0
backcall==0.2.0
base58==2.1.1
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
construct==2.10.67
construct-typing==0.5.2
decorator==5.1.1
dill==0.3.4
executing==0.8.3
Flask==2.0.3
Flask-Cors==3.0.10
h11==0.12.0
httpcore==0.13.7
httpx==0.18.2
humanize==4.0.0
idna==3.3
ipython==8.1.1
itsdangerous==2.1.1
jedi==0.18.1
Jinja2==3.0.3
jsonrpcclient==4.0.2
jsonrpcserver==5.0.6
jsons==1.6.1
jsonschema==3.2.0
mango-explorer==3.4.3
MarkupSafe==2.1.0
matplotlib-inline==0.1.3
multiprocess==0.70.12.2
numpy==1.22.3
OSlash==0.6.3
pandas==1.4.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.28
psycopg-binary==3.0.11
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.11.2
PyNaCl==1.5.0
pyrsistent==0.18.1
pyserum==0.5.0a0
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
rfc3986==1.5.0
Rx==3.2.0
rxpy-backpressure==1.0.0
six==1.16.0
sniffio==1.2.0
solana==0.21.0
stack-data==0.2.0
traitlets==5.1.1
types-cachetools==4.2.10
typing-extensions==3.10.0.2
typish==1.9.3
urllib3==1.26.8
wcwidth==0.2.5
websocket-client==1.3.1
websockets==10.2
Werkzeug==2.0.3
zstandard==0.17.0

matplotlib~=3.5.1